Typescript strict type options

+++ src/client/app/actions/auth.tsx
@@ -8,6 +8,7 @@ import { receiveUserDetails } from 'actions/currentUser';
 import { DIALOG_LOADER_ID } from 'client-utils/constants';
 import { getUserDetails } from 'client-utils/methods';
 import { getAccessTokenClaims } from 'global-utils/methods';
+import { getAccessToken } from 'selectors';
 import {
   IAppState,
   AuthActionTypes,
@@ -16,7 +17,8 @@ import {
   IShowKeepMeLoggedModal,
   IReauthenticateSuccess,
   ISetAccessToken,
-  Toast
+  Toast,
+  ThunkResult
 } from 'types';
 
 import {
@@ -51,7 +53,7 @@ export const setAccessToken = (accessToken: string): ISetAccessToken => ({
   accessToken
 });
 
-export const handleAuthError = (dispatch, isLogin: boolean) => (error) => {
+export const handleAuthError = (dispatch: any, isLogin: boolean) => (error: any) => {
   const genericMessage = isLogin ? USER_LOGIN_ERROR : USER_LOGOUT_ERROR;
   const errorMessage = error.response.data.message;
   console.error(error);
@@ -60,7 +62,7 @@ export const handleAuthError = (dispatch, isLogin: boolean) => (error) => {
   dispatch(logout());
 };
 
-export const login = (credentials = {username: 'admin', password: 'admin'}) => dispatch => {
+export const login = (credentials = {username: 'admin', password: 'admin'}): ThunkResult<Promise<void>> => dispatch => {
   dispatch(startLoading(DIALOG_LOADER_ID));
   return http.post('/api/users/login', credentials)
     .then(response => response.data)
@@ -82,11 +84,14 @@ export const login = (credentials = {username: 'admin', password: 'admin'}) => d
     .catch(handleAuthError(dispatch, true));
 };
 
-export const reauthenticateUser = (displayToast = false) => (dispatch, getState) => {
+export const reauthenticateUser = (displayToast = false): ThunkResult<Promise<void> | null> => (dispatch, getState) => {
   dispatch(startLoading(DIALOG_LOADER_ID));
-  const state: IAppState = getState();
-  const oldAccessToken = state.auth.accessToken;
+  const state = getState();
+  const oldAccessToken = getAccessToken(state);
   const refreshToken = localStorage.getItem('refreshToken');
+
+  if (!oldAccessToken) return null;
+
   const dataToSend = {...getAccessTokenClaims(oldAccessToken), refreshToken };
   return http.post('/api/users/reauthenticate', dataToSend)
     .then(response => response.data)
@@ -104,9 +109,12 @@ export const reauthenticateUser = (displayToast = false) => (dispatch, getState)
     .catch(handleAuthError(dispatch, true));
 };
 
-export const logout = () => (dispatch, getState) => {
-  const state: IAppState = getState();
-  const accessToken = state.auth.accessToken;
+export const logout = (): ThunkResult<Promise<void> | null> => (dispatch, getState) => {
+  const state = getState();
+  const accessToken = getAccessToken(state);
+
+  if (!accessToken) return null;
+
   const { userId } = getAccessTokenClaims(accessToken);
   return http.delete(`/api/users/logout/${userId}`)
     .then(response => response.data)
@@ -115,7 +123,7 @@ export const logout = () => (dispatch, getState) => {
       localStorage.removeItem('refreshToken');
       dispatch(logoutSuccess());
       dispatch(showToast(Toast.success, USER_LOGOUT_SUCCESS));
-      dispatch(receiveUserDetails(null));
+      dispatch(receiveUserDetails({ userDetails: null }));
     })
     .catch(handleAuthError(dispatch, false));
 };
